config: change the structure of the web app to a module
